requirements.txt
Discord.py
